Additional logging and move isScottish to RPC checker
