+++ src/main/java/uk/gov/hmcts/reform/sscs/services/ccd/UpdateCcdRpc.java
@@ -1,5 +1,7 @@
 package uk.gov.hmcts.reform.sscs.services.ccd;
 
+import static java.util.Objects.isNull;
+
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.domain.*;
@@ -8,6 +10,7 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.*;
 @Service
 class UpdateCcdRpc {
     boolean updateCcdRpc(SscsCaseData gapsCaseData, SscsCaseData existingCcdCaseData) {
+
         boolean rpcUpdated = false;
 
         if (gapsCaseData == null
@@ -21,11 +24,16 @@ class UpdateCcdRpc {
 
         if (hasRpcChanged(gapsRpc, existingRpc)) {
 
-            log.info("RPC has changed from {} to {} for case {}", toName(existingRpc), toName(existingRpc),
+            log.info("RPC has changed from {} to {} for case {}", toName(existingRpc), toName(gapsRpc),
                 existingCcdCaseData.getCcdCaseId());
 
             existingCcdCaseData.setRegionalProcessingCenter(gapsRpc);
             existingCcdCaseData.setRegion(gapsRpc.getName());
+
+            String isScottishCase = isScottishCase(gapsRpc, existingCcdCaseData);
+
+            existingCcdCaseData.setIsScottishCase(isScottishCase);
+
             rpcUpdated = true;
         } else {
             log.info("RPC has not changed for case {} . RPC =  {}", existingCcdCaseData.getCcdCaseId(),
@@ -48,6 +56,19 @@ class UpdateCcdRpc {
             || gapsRpc.getAddress1() == null
             || !existingRpc.getName().equals(gapsRpc.getName())
             || !existingRpc.getAddress1().equals(gapsRpc.getAddress1());
+    }
+
+    public static String isScottishCase(RegionalProcessingCenter rpc, SscsCaseData caseData) {
+
+        if (isNull(rpc) || isNull(rpc.getName())) {
+            log.info("Setting isScottishCase field to No for empty RPC for case " + caseData.getCcdCaseId());
+            return "No";
+        } else {
+            String isScotCase = rpc.getName().equalsIgnoreCase("GLASGOW") ? "Yes" : "No";
+            log.info("Setting isScottishCase field to " + isScotCase + " for RPC " + rpc.getName() + " for case "
+                + caseData.getCcdCaseId());
+            return isScotCase;
+        }
     }
 
 }
